tests: display: reuse driver display sample code for display_check
display_check is sharing same code as sample driver display.
Avoid duplicating the code and directly reuse the sample code for
the test.
