+++ samples/drivers/display/CMakeLists.txt
@@ -4,4 +4,4 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(display)
 
-target_sources(app PRIVATE src/main.c)
+target_sources(app PRIVATE src/main.c src/display.c)
