--- tests/drivers/display/display_check/boards/mimxrt1170_evk_mimxrt1176_cm7.conf
@@ -1,10 +1,0 @@
-#
-# Copyright 2022, NXP
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# Sample will allocate buffer equal to: (panelwidth / 8) * (panelwidth / 4) * pixel depth. For a
-# 1280x720 display in a 32-bpp format (e.g. ARGB8888), this is (720 / 8) * (720 / 4) * 4 = 64800
-# bytes. We include 128 bytes of padding for kernel heap structures
-CONFIG_HEAP_MEM_POOL_SIZE=64928
