+++ tests/drivers/display/display_check/boards/mimxrt700_evk_mimxrt798s_cm33_cpu0.conf
@@ -1,0 +1,8 @@
+#
+# Copyright 2025 NXP
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+CONFIG_SAMPLE_BUFFER_ADDR_ALIGN=64
+CONFIG_SAMPLE_PITCH_ALIGN=64
