--- tests/drivers/display/display_check/boards/stm32n6570_dk.conf
@@ -1,6 +1,0 @@
-# Copyright (c) 2025 ST Microelectronics
-# SPDX-License-Identifier: Apache-2.0
-
-# On STM32N6 DK board, display being larger, it requires
-# more memory to draw squares on the display (40K)
-CONFIG_HEAP_MEM_POOL_SIZE=65536
