+++ tests/drivers/display/display_check/boards/stm32u5g9j_dk2.conf
@@ -4,7 +4,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-CONFIG_HEAP_MEM_POOL_SIZE=172840
 CONFIG_IDLE_STACK_SIZE=8192
 CONFIG_PRIVILEGED_STACK_SIZE=8192
 CONFIG_MAIN_STACK_SIZE=4096
