+++ tests/drivers/display/display_check/prj.conf
@@ -1,7 +1,6 @@
 # Copyright 2025 NXP
 # SPDX-License-Identifier: Apache-2.0
 
-CONFIG_HEAP_MEM_POOL_SIZE=16384
 CONFIG_LOG=y
 CONFIG_DISPLAY=y
 CONFIG_ZTEST=y
